requirements/base.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
filelock==3.12.1
flake8==6.0.0
frozenlist==1.3.3
identify==2.5.24
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
jsonschema==4.17.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pre-commit==3.3.2
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
tomli==2.0.1
tqdm==4.65.0
urllib3==2.0.3
virtualenv==20.23.0
yarl==1.9.2
